Add code coverage processing and sanity checking
In addition to being able to generate code coverage information using
gcov, llvm-cov, and gcovr using the —coverage or —html-coverage
options, the runTests.py script now checks to make sure it is in the
correct directory and that Mongoose has already been built. If not, it
exits and prints an error.