Add additional constructor for making a graph from a matrix
Previously, only one constructor existed for creating a graph from a matrix. That constructor would play it safe and never free any of its arrays, as it was assumed the user would be passing in the matrix. However, internally, we can create a graph from a matrix that we obtained from reading a file. We need make sure that these arrays are freed.